+++ XRViewer/Resources/XRViewer-Bridging-Header.h
@@ -2,7 +2,6 @@
 //  Use this file to import your target's public headers that you would like to expose to Swift.
 //
 
-#import "OverlayHeader.h"
 #import "ARKHelper.h"
 #import "Prefix.h"
 #import "ARKController.h"
@@ -13,3 +12,5 @@ typedef NS_ENUM(NSInteger, ResetTrackingOption) {
     SaveWorldMap,
     LoadSavedWorldMap
 };
+
+typedef void (^HotAction)(BOOL); // long